Extract functions required by SingleMediaSelection as well
