+++ src/Sulu/Bundle/MediaBundle/Resources/js/containers/MultiMediaSelection/types.js
@@ -1,16 +1,7 @@
 // @flow
+import type {DisplayOption} from '../../types';
+
 export type Value = {
     displayOption: ?DisplayOption,
     ids: Array<number>,
 };
-
-export type DisplayOption =
-    | 'leftTop'
-    | 'top'
-    | 'rightTop'
-    | 'left'
-    | 'middle'
-    | 'right'
-    | 'leftBottom'
-    | 'bottom'
-    | 'rightBottom';
