+++ src/Sulu/Bundle/MediaBundle/Resources/js/utils/MediaSelectionHelper/index.js
@@ -1,0 +1,12 @@
+// @flow
+import convertDisplayOptionsFromParams from './convertDisplayOptionsFromParams';
+import getIconForDisplayOption from './getIconForDisplayOption';
+import getTranslationForDisplayOption from './getTranslationForDisplayOption';
+import validateDisplayOption from './validateDisplayOption';
+
+export {
+    convertDisplayOptionsFromParams,
+    getIconForDisplayOption,
+    getTranslationForDisplayOption,
+    validateDisplayOption,
+};
